Refactor IPhysicalMemory to use Addr (uint) for addresses and uint for sizes
- Change Size property from int to uint
- Change Slice/ReadOnlySlice offset and length from int to uint
- Change SlicePage/ReadOnlySlicePage pageIndex and pageSize from int to uint
- Change PageCount return type and pageSize from int to uint
- Change WritePhysical/WriteBytePhysical address from int to Addr (uint)
- Update PhysicalMemory implementation to match
- Update RamTarget and RomTarget Size properties to uint
- Remove tests for negative values (no longer applicable with unsigned types)

+++ src/BadMango.Emulator.Bus/RamTarget.cs
@@ -56,7 +56,7 @@ public sealed class RamTarget : IBusTarget
     /// <summary>
     /// Gets the size of the RAM in bytes.
     /// </summary>
-    public int Size => memory.Length;
+    public uint Size => (uint)memory.Length;
 
     /// <inheritdoc />
     public byte Read8(Addr physicalAddress, in BusAccess access)
